requirements.txt
aiofiles==23.1.0
asgiref==3.6.0
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.2
django-cors-headers==3.14.0
django-jazzmin==2.6.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
httptools==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.0
Pillow==9.5.0
platformdirs==2.6.2
pycodestyle==2.10.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
sanic==22.9.1
Sanic-Cors==2.2.0
sanic-routing==22.8.0
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
virtualenv==20.17.1
websockets==10.4
Werkzeug==2.2.2
